+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Make `note` property non-nullable in Mastodon API `Account` object.
 - Renamed `apresolver` well-known endpoint to `apgateway`.
 - Don't write warning to log if actor's public key changes.
+- Increase importer limit from 50 items to 500.
 
 ### Fixed